+++ source/InputSmart.js
@@ -14,13 +14,14 @@ export function createInput(defaultMetadata)
 	function InputSmart({
 		country,
 		international,
+		withCountryCallingCode,
 		metadata,
 		...rest
 	}, ref) {
 		const format = useCallback((value) => {
 			// "As you type" formatter.
 			const formatter = new AsYouType(country, metadata)
-			const prefix = getInputValuePrefix(country, international, metadata)
+			const prefix = getInputValuePrefix(country, international, withCountryCallingCode, metadata)
 			// Format the number.
 			let text = formatter.input(prefix + value)
 			let template = formatter.getTemplate()
@@ -71,6 +72,14 @@ export function createInput(defaultMetadata)
 		 * (without "country calling code" `+1`).
 		 */
 		international: PropTypes.bool,
+
+		/**
+		 * If `country` and `international` properties are set,
+		 * then by default it won't include "country calling code" in the input field.
+		 * To change that, pass `withCountryCallingCode` property,
+		 * and it will include "country calling code" in the input field.
+		 */
+		withCountryCallingCode: PropTypes.bool,
 
 		/**
 		 * `libphonenumber-js` metadata.